Initial implementation of an Asset Manager.
The AssetManager tracks loaders for various formats and provides a centralized
location for loading assets. This allows us to cache assets to be provided
at anytime and anyplace in the code. Assets are organized into a stack that
can partially purge all assets associated with a state/level without touching
assets needed elsewhere.

TODO: Not fully tested. Support for various loaders still needs to be completed.